First Bit of Code
Added code to read and load all data from the given file. Stored in an appropriate Array List. Made sure to add the given data to my netbeans project. Added .shuffle to randomise the order of the data. Added a buffered reader with a try/catch to read each line of the file.